Comments from https://github.com/google/ExoPlayer/pull/2963#discussion_r122669328 applied
