+++ library/core/src/androidTest/java/com/google/android/exoplayer2/upstream/AndroidDataSourceConstants.java
@@ -1,0 +1,8 @@
+package com.google.android.exoplayer2.upstream;
+
+final class AndroidDataSourceConstants {
+  static final long SAMPLE_MP4_BYTES = 101597;
+  static final String SAMPLE_MP4_PATH = "/mp4/sample.mp4";
+
+  private AndroidDataSourceConstants() {}
+}
